# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for StreamShort, a short-form video streaming platform.",
        "title": "StreamShort API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/creators": {
            "get": {
                "description": "List all creators.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List Creators",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CreatorProfile"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/creators/{id}": {
            "delete": {
                "description": "Delete a creator profile.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete Creator Profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/creators/onboard": {
            "post": {
                "description": "Register a new creator profile for the user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "Onboard as Creator",
                "parameters": [
                    {
                        "description": "Onboard Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatorOnboardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreatorProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/creators/profile": {
            "get": {
                "description": "Get the creator profile for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "Get Own Creator Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreatorProfile"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the creator profile for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "Update Creator Profile",
                "parameters": [
                    {
                        "description": "Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatorOnboardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreatorProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/creators/{id}/dashboard": {
            "get": {
                "description": "Get dashboard analytics and stats for a creator.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "Get Creator Dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatorDashboardResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/creators/{id}/follow": {
            "post": {
                "description": "Follow a creator by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "Follow Creator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Unfollow a creator by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "Unfollow Creator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/creators/{id}/following": {
            "get": {
                "description": "Check if the user is following a specific creator.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "Check Following Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/me/following": {
            "get": {
                "description": "List creators followed by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator"
                ],
                "summary": "List Following",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "handlers.CreatorDashboardResponse": {
            "type": "object",
            "properties": {
                "earnings": {
                    "type": "number"
                },
                "follower_count": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                },
                "watch_time_seconds": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.CreatorOnboardRequest": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "kyc_document_s3_path": {
                    "type": "string"
                }
            }
        },
...
        "models.CreatorProfile": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "display_name": {
                    "type": "string"
                },
                "follower_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "kyc_document_s3_path": {
                    "type": "string"
                },
                "kyc_status": {
                    "type": "string"
                },
                "payout_details": {
                    "$ref": "#/definitions/models.PayoutDetails"
                },
                "rating": {
                    "type": "number"
                },
                "series_count": {
                    "description": "Derived fields (not stored)",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.PayoutDetails": {
            "type": "object",
            "properties": {
                "account_holder": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "ifsc_code": {
                    "type": "string"
                },
                "razorpay_contact_id": {
                    "type": "string"
                },
                "razorpay_fund_account_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "upi_id": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creator_profile": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CreatorProfile"
                        }
                    ]
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "'user' or 'admin'",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "description": "Only for admins (nullable to allow NULL for non-admins)",
                    "type": "string"
                }
            }
        }
    }
}