# handlers/creator.go
package handlers

import (
	"encoding/json"
	"net/http"
	"strconv"
	"time"

	"streamshort/models"

	"github.com/gorilla/mux"
	"gorm.io/gorm"
)

type CreatorHandler struct {
	db *gorm.DB
}

func NewCreatorHandler(db *gorm.DB) *CreatorHandler {
	return &CreatorHandler{db: db}
}

// Request/Response structs matching OpenAPI schema
type CreatorOnboardRequest struct {
	DisplayName     string `json:"display_name"`
	Bio             string `json:"bio"`
	KYCDocumentPath string `json:"kyc_document_s3_path"`
}

type CreatorDashboardResponse struct {
	Views            int64   `json:"views"`
	WatchTimeSeconds int64   `json:"watch_time_seconds"`
	Earnings         float64 `json:"earnings"`
	FollowerCount    int64   `json:"follower_count"`
}

// Creator onboarding endpoint
// OnboardCreator godoc
// @Summary Onboard as Creator
// @Description Register a new creator profile for the user.
// @Tags creator
// @Accept  json
// @Produce  json
// @Param   request  body     CreatorOnboardRequest  true  "Onboard Request"
// @Success 201 {object} models.CreatorProfile
// @Failure 400 {string} string "Bad Request"
// @Failure 409 {string} string "Conflict"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/creators/onboard [post]
func (h *CreatorHandler) OnboardCreator(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context (set by auth middleware)
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	var req CreatorOnboardRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// Validate required fields
	if req.DisplayName == "" {
		http.Error(w, "Display name is required", http.StatusBadRequest)
		return
	}

	if req.KYCDocumentPath == "" {
		http.Error(w, "KYC document path is required", http.StatusBadRequest)
		return
	}

	// Check if user already has a creator profile
	var existingProfile models.CreatorProfile
	if err := h.db.Where("user_id = ?", userID).First(&existingProfile).Error; err == nil {
		http.Error(w, "Creator profile already exists for this user", http.StatusConflict)
		return
	} else if err != gorm.ErrRecordNotFound {
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	// Create creator profile
	creatorProfile := models.CreatorProfile{
		UserID:          userID,
		DisplayName:     req.DisplayName,
		Bio:             req.Bio,
		KYCDocumentPath: req.KYCDocumentPath,
		KYCStatus:       "pending",
	}

	if err := h.db.Create(&creatorProfile).Error; err != nil {
		http.Error(w, "Failed to create creator profile", http.StatusInternalServerError)
		return
	}

	// Return the created profile
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(creatorProfile)
}

// Creator dashboard endpoint
// GetCreatorDashboard godoc
// @Summary Get Creator Dashboard
// @Description Get dashboard analytics and stats for a creator.
// @Tags creator
// @Produce  json
// @Param   id   path     string  true  "Creator ID"
// @Success 200 {object} CreatorDashboardResponse
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/creators/{id}/dashboard [get]
func (h *CreatorHandler) GetCreatorDashboard(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	// Get creator ID from URL path
	vars := mux.Vars(r)
	creatorID := vars["id"]

	// Verify that the user is accessing their own dashboard
	var creatorProfile models.CreatorProfile
	if err := h.db.Where("id = ? AND user_id = ?", creatorID, userID).First(&creatorProfile).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			http.Error(w, "Creator profile not found or access denied", http.StatusNotFound)
			return
		}
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	// Get analytics for the last 30 days
	thirtyDaysAgo := time.Now().AddDate(0, 0, -30)

	var analytics []models.CreatorAnalytics
	if err := h.db.Where("creator_id = ? AND date >= ?", creatorID, thirtyDaysAgo).
		Find(&analytics).Error; err != nil {
		http.Error(w, "Failed to fetch analytics", http.StatusInternalServerError)
		return
	}

	// Aggregate analytics (views and watch time from creator_analytics)
	var totalViews int64
	var totalWatchTime int64

	for _, analytic := range analytics {
		totalViews += analytic.Views
		totalWatchTime += analytic.WatchTimeSeconds
	}

	// Get earnings from creator_earnings table (actual earnings from subscriptions, purchases, ad revenue)
	var totalEarnings float64
	if err := h.db.Model(&models.CreatorEarnings{}).
		Where("creator_id = ? AND status != 'cancelled'", creatorID).
		Select("COALESCE(SUM(amount), 0)").
		Scan(&totalEarnings).Error; err != nil {
		// Log error but continue with 0 earnings
		totalEarnings = 0
	}

	response := CreatorDashboardResponse{
		Views:            totalViews,
		WatchTimeSeconds: totalWatchTime,
		Earnings:         totalEarnings,
		FollowerCount:    creatorProfile.FollowerCount,
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// Get creator profile endpoint
// GetCreatorProfile godoc
// @Summary Get Own Creator Profile
// @Description Get the creator profile for the authenticated user.
// @Tags creator
// @Produce  json
// @Success 200 {object} models.CreatorProfile
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/creators/profile [get]
func (h *CreatorHandler) GetCreatorProfile(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	// Get creator profile for the authenticated user
	var creatorProfile models.CreatorProfile
	if err := h.db.Where("user_id = ?", userID).First(&creatorProfile).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			http.Error(w, "Creator profile not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	// Populate derived field: series_count
	var seriesCount int64
	if err := h.db.Model(&models.Series{}).Where("creator_id = ?", creatorProfile.ID).Count(&seriesCount).Error; err == nil {
		creatorProfile.SeriesCount = seriesCount
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(creatorProfile)
}

// Follow a creator
// FollowCreator godoc
// @Summary Follow Creator
// @Description Follow a creator by ID.
// @Tags creator
// @Produce  json
// @Param   id   path     string  true  "Creator ID"
// @Success 200 {object} map[string]string
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/creators/{id}/follow [post]
func (h *CreatorHandler) FollowCreator(w http.ResponseWriter, r *http.Request) {
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}
	vars := mux.Vars(r)
	creatorID := vars["id"]

	var creator models.CreatorProfile
	if err := h.db.Where("id = ?", creatorID).First(&creator).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			http.Error(w, "Creator not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	follow := models.CreatorFollow{FollowerID: userID, CreatorID: creatorID}
	tx := h.db.Where("follower_id = ? AND creator_id = ?", userID, creatorID).FirstOrCreate(&follow)
	if tx.Error != nil {
		http.Error(w, "Failed to follow creator", http.StatusInternalServerError)
		return
	}
	if tx.RowsAffected == 1 {
		_ = h.db.Model(&models.CreatorProfile{}).
			Where("id = ?", creatorID).
			UpdateColumn("follower_count", gorm.Expr("follower_count + 1")).Error
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "followed"})
}

// Unfollow a creator
// UnfollowCreator godoc
// @Summary Unfollow Creator
// @Description Unfollow a creator by ID.
// @Tags creator
// @Produce  json
// @Param   id   path     string  true  "Creator ID"
// @Success 200 {object} map[string]string
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/creators/{id}/follow [delete]
func (h *CreatorHandler) UnfollowCreator(w http.ResponseWriter, r *http.Request) {
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}
	vars := mux.Vars(r)
	creatorID := vars["id"]

	txDel := h.db.Where("follower_id = ? AND creator_id = ?", userID, creatorID).Unscoped().Delete(&models.CreatorFollow{})
	if txDel.Error != nil {
		http.Error(w, "Failed to unfollow creator", http.StatusInternalServerError)
		return
	}
	if txDel.RowsAffected > 0 {
		_ = h.db.Model(&models.CreatorProfile{}).
			Where("id = ? AND follower_count > 0", creatorID).
			UpdateColumn("follower_count", gorm.Expr("follower_count - 1")).Error
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "unfollowed"})
}

// List creators the user follows
// ListFollowing godoc
// @Summary List Following
// @Description List creators followed by the authenticated user.
// @Tags creator
// @Produce  json
// @Param   page         query    int     false  "Page number"
// @Param   limit        query    int     false  "Results per page"
// @Success 200 {object} map[string]interface{}
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/me/following [get]
func (h *CreatorHandler) ListFollowing(w http.ResponseWriter, r *http.Request) {
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}
	page := 1
	limit := 20
	if p, err := strconv.Atoi(r.URL.Query().Get("page")); err == nil && p > 0 {
		page = p
	}
	if l, err := strconv.Atoi(r.URL.Query().Get("limit")); err == nil && l > 0 && l <= 100 {
		limit = l
	}

	var follows []models.CreatorFollow
	if err := h.db.Where("follower_id = ?", userID).Offset((page - 1) * limit).Limit(limit).Order("created_at DESC").Find(&follows).Error; err != nil {
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}
	creatorIDs := make([]string, 0, len(follows))
	for _, f := range follows {
		creatorIDs = append(creatorIDs, f.CreatorID)
	}

	var creators []models.CreatorProfile
	if len(creatorIDs) > 0 {
		if err := h.db.Where("id IN ?", creatorIDs).Find(&creators).Error; err != nil {
			http.Error(w, "Database error", http.StatusInternalServerError)
			return
		}
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{"total": len(creators), "items": creators})
}

// Check if current user follows a creator
// IsFollowing godoc
// @Summary Check Following Status
// @Description Check if the user is following a specific creator.
// @Tags creator
// @Produce  json
// @Param   id   path     string  true  "Creator ID"
// @Success 200 {object} map[string]bool
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/creators/{id}/following [get]
func (h *CreatorHandler) IsFollowing(w http.ResponseWriter, r *http.Request) {
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}
	vars := mux.Vars(r)
	creatorID := vars["id"]

	var count int64
	if err := h.db.Model(&models.CreatorFollow{}).Where("follower_id = ? AND creator_id = ?", userID, creatorID).Count(&count).Error; err != nil {
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]bool{"following": count > 0})
}

// Update creator profile endpoint
// UpdateCreatorProfile godoc
// @Summary Update Creator Profile
// @Description Update the creator profile for the authenticated user.
// @Tags creator
// @Accept  json
// @Produce  json
// @Param   request  body     CreatorOnboardRequest  true  "Update Request"
// @Success 200 {object} models.CreatorProfile
// @Failure 400 {string} string "Bad Request"
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/creators/profile [put]
func (h *CreatorHandler) UpdateCreatorProfile(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	var req CreatorOnboardRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// Get existing creator profile
	var creatorProfile models.CreatorProfile
	if err := h.db.Where("user_id = ?", userID).First(&creatorProfile).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			http.Error(w, "Creator profile not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	// Update fields
	if req.DisplayName != "" {
		creatorProfile.DisplayName = req.DisplayName
	}
	if req.Bio != "" {
		creatorProfile.Bio = req.Bio
	}
	if req.KYCDocumentPath != "" {
		creatorProfile.KYCDocumentPath = req.KYCDocumentPath
		// Reset KYC status to pending when document is updated
		creatorProfile.KYCStatus = "pending"
	}

	// Save changes
	if err := h.db.Save(&creatorProfile).Error; err != nil {
		http.Error(w, "Failed to update creator profile", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(creatorProfile)
}